Allow pattern matching in ResultMatcher
A gotcha that has bitten my team several times is that you cannot
pattern match against the wrapped value of a result in the
ResultMatcher.

Since the [basic matcher documentation][1] shows an example where you
can pattern match against the patterns that you specify in your case
block, this is easy to assume that it's a built-in for the
ResultMatcher.

Adding pattern matching is a simple change that makes the whole of the
ResultMatcher more flexible and more in-line with the examples in the
rest of the documentation.

[1]: https://dry-rb.org/gems/dry-matcher/